% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cer_func.R
\name{cer_func}
\alias{cer_func}
\title{Conditional Error Rate Calculator}
\usage{
cer_func(ind, predictObject, newdatagroup)
}
\arguments{
\item{ind}{logical on whether the new data ahd a name column}

\item{predictObject}{prediction object}

\item{newdatagroup}{new data group}
}
\description{
Conditional Error Rate Calculator
}
\keyword{internal}
