% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_reduceOutput.R
\name{check_reduceOutput}
\alias{check_reduceOutput}
\title{Check The Reduce Output and Output a matrix, data.frame, or grouped_df}
\usage{
check_reduceOutput(dat, output, group)
}
\arguments{
\item{dat}{data}

\item{output}{output type either matrix, data.frame, or grouped_df}

\item{group}{grouping variable name}
}
\description{
Check The Reduce Output and Output a matrix, data.frame, or grouped_df
}
\keyword{internal}
