% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_targetDimType.R
\name{check_targetDimType}
\alias{check_targetDimType}
\title{Check to verify the targetDim Type and Produce a target dimension}
\usage{
check_targetDimType(targetDim, x, ...)
}
\arguments{
\item{targetDim}{function or numeric for producing the target dimension}

\item{x}{a list of matrices}

\item{...}{other options passsed to targetDim functions}
}
\description{
Check to verify the targetDim Type and Produce a target dimension
}
