% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Classification function}
\usage{
classify(x, ...)
}
\arguments{
\item{x}{data set}
}
\description{
Given a dimension reduced data set a classifier is built using selected discrimination function.
}

