% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConcatScatterBlock.R
\name{DataConcatScatterBlock}
\alias{DataConcatScatterBlock}
\title{Data Concatenated and Scatter Matrix}
\usage{
DataConcatScatterBlock(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{}
}
\description{
Data Concatenated and Scatter Matrix
}
\examples{
DataConcatScatterBlock(iris, group = Species, targetDim = 1)
}

