% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{dataDftoMatrixDim}
\alias{dataDftoMatrixDim}
\title{Turn no a Tidy data frame into data matrix (helper function)}
\usage{
dataDftoMatrixDim(data, group, method, ..., .dots)
}
\arguments{
\item{data}{not a tidy dataframe}
}
\description{
Turn no a Tidy data frame into data matrix (helper function)
}
\keyword{internal}

