% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrimFunc.R
\name{discrimFunc}
\alias{discrimFunc}
\title{Discrimination Function Classifier}
\usage{
discrimFunc(dat, discrim, ...)
}
\arguments{
\item{dat}{data}

\item{group}{grouping variable}
}
\description{
Discrimination Function Classifier
}
\keyword{internal}
