% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HighDimSY.R
\name{HighDimSY}
\alias{HighDimSY}
\alias{HighDimSY.data.frame}
\alias{HighDimSY.grouped_df}
\alias{HighDimSY.resample}
\alias{HighDimSY.matrix}
\title{High Dimensional SY}
\usage{
HighDimSY(x, ...)

\method{HighDimSY}{data.frame}(x, group, targetDim, ...)

\method{HighDimSY}{grouped_df}(x, targetDim, ...)

\method{HighDimSY}{resample}(x, targetDim, ...)

\method{HighDimSY}{matrix}(..., targetDim, svdMethod = svd)
}
\arguments{
\item{x}{data}

\item{...}{other options}

\item{group}{grouping variable}

\item{targetDim}{target dimension to reduce the data to}

\item{svdMethod}{svd function used for dimesion reduction by default 
svd in base is used}
}
\value{
list of reduced data, projection matrix, 
         group variable, discrimination function, 
         m matrix.
}
\description{
High Dimensional SY
}
\examples{
HighDimSY(iris, group = Species, targetDim = 1)
}
