% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listMatToDataframe.R
\name{listMatToGroupedDataframe}
\alias{listMatToGroupedDataframe}
\title{List of Matrices to Grouped Data Frame}
\usage{
listMatToGroupedDataframe(x, group)
}
\description{
List of Matrices to Grouped Data Frame
}
\keyword{internal}
