% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConcatScatterBlock.R
\name{matBlocks}
\alias{matBlocks}
\title{Helper for Data Concat Scatter Block}
\usage{
matBlocks(x, rows, columns)
}
\arguments{
\item{x}{data}

\item{rows}{number of rows}

\item{columns}{number of columns}
}
\description{
Helper for Data Concat Scatter Block
}
\keyword{internal}

