% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSIR.R
\name{MSIR}
\alias{MSIR}
\title{Modified Sliced Inverse Regression}
\usage{
MSIR(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{}
}
\description{
Modified Sliced Inverse Regression
}
\examples{
MSIR(mcSamples(c(0,0,0), diag(1, 3), 10, 3, matrix = FALSE, tidy = TRUE), group = population, targetDim = 1, variables = variable, samples = samples, value = value, tidy = TRUE)
}

