% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Original.R
\name{Original}
\alias{Original}
\title{Original Data}
\usage{
Original(x, group, targetDim, svdMethod)
}
\arguments{
\item{x}{data}

\item{group}{grouping variable}

\item{targetDim}{target dimension to reduce the data to}

\item{svdMethod}{svd function used for dimesion reduction by default 
svd in base is used}
}
\description{
Original Data
}
\keyword{internal}
