% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcSamples.R
\name{pop_lists}
\alias{pop_lists}
\title{List population parameters}
\usage{
pop_lists(meanVec, covMat, samples, pops)
}
\arguments{
\item{meanVec}{vector of means}

\item{covMat}{covariance matrices}

\item{samples}{number of samples}

\item{pops}{number of populations}
}
\description{
List population parameters
}
\keyword{internal}
