% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_func.R
\name{projection_func}
\alias{projection_func}
\title{Projection Function}
\usage{
projection_func(x, proj)
}
\arguments{
\item{x}{data matrix}

\item{proj}{projection matrix}
}
\description{
Projection Function
}
