% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projection_matrix.R
\name{projection_matrix}
\alias{projection_matrix}
\alias{projection_matrix.reduced}
\alias{projection_matrix.list}
\alias{projection_matrix.default}
\title{Projection Matrix}
\usage{
projection_matrix(x, ...)

\method{projection_matrix}{reduced}(x, ...)

\method{projection_matrix}{list}(x, method = SYS, ...)

\method{projection_matrix}{default}(x, method = SYS, ..., group)
}
\arguments{
\item{x}{data or reduced class object}

\item{...}{options past to linear dimension reduction methods}

\item{method}{linear dimension reduction methods}

\item{group}{the grouping variable in the data}
}
\description{
Given a linear dimension reduction method projection_matrix parses the data
and uses the specific linear dimension reduction to return the projection
matrix. The projection_matrix function should be used if the user wishes just
to produce the projection matrix as opposed to the data reduction matrix.
}
