% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movetoCovEstR.R
\name{S_B}
\alias{S_B}
\title{Scatter Matrix Between}
\usage{
S_B(prior, xbar)
}
\arguments{
\item{prior}{prior based on sample sizes of groups}

\item{xbar}{sample means}

\item{mu}{overall mean}
}
\description{
Scatter Matrix Between
}
\keyword{internal}
