% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S_w.R
\name{S_W}
\alias{S_W}
\title{Within group covarinace}
\usage{
S_W(prior, matrix_ls)
}
\arguments{
\item{prior}{prior based on group samples}

\item{matrix_ls}{group sample covarinace matrices}
}
\description{
Within group covarinace
}
\keyword{internal}
