% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterMat.R
\name{scatterMat}
\alias{scatterMat}
\title{Sum of Squares (helper function)}
\usage{
scatterMat(matrix)
}
\arguments{
\item{matrix}{data matrix}
}
\description{
Sum of Squares (helper function)
}
\keyword{internal}
