% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDiff.R
\name{SDiff}
\alias{SDiff}
\alias{SDiff.resample}
\title{Sample Covariance Differences Concatenated}
\usage{
SDiff(x, ...)

\method{SDiff}{resample}(x, targetDim, ...)
}
\arguments{
\item{x}{data}

\item{...}{other options}

\item{targetDim}{target dimension to reduce the data to}
}
\description{
Sample Covariance Differences Concatenated
}
\keyword{internal}
