% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slow_learn.R
\name{slow_learn}
\alias{slow_learn}
\title{Slow Learning Algorithm}
\usage{
slow_learn(x, ...)
}
\arguments{
\item{...}{}
}
\description{
Slow Learning Algorithm
}
\examples{
slow_learn(iris, group = Species, loss = conditional_loss, lossValue = .1, method = SYS)
}

