% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SYS.R
\name{SYS}
\alias{SYS}
\alias{SYS.data.frame}
\alias{SYS.grouped_df}
\alias{SYS.resample}
\alias{SYS.matrix}
\title{SYS}
\usage{
SYS(x, ...)

\method{SYS}{data.frame}(x, group, targetDim, ...)

\method{SYS}{grouped_df}(x, targetDim, ...)

\method{SYS}{resample}(x, targetDim, ...)

\method{SYS}{matrix}(..., group, targetDim, svdMethod = svd,
  covest = Haff_shrinkage)
}
\arguments{
\item{x}{data}

\item{...}{other options}

\item{group}{grouping variable}

\item{targetDim}{target dimension to reduce the data to}

\item{svdMethod}{svd function used for dimesion reduction by default 
svd in base is used}

\item{covest}{covariance estimator used forming M matrix}
}
\value{
list of reduced data, projection matrix, 
         group variable, discrimination function, 
         m matrix.
}
\description{
SYS
}
\examples{
SYS(iris, group = Species, targetDim = 1)
}
