% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{tidyDataDftoMatrixDim}
\alias{tidyDataDftoMatrixDim}
\title{Turn Tidy data frame into data matrix (helper function)}
\usage{
tidyDataDftoMatrixDim(data, group, variables, samples, value, test, targetDim,
  svdMethod)
}
\arguments{
\item{data}{tidy dataframe}
}
\description{
Turn Tidy data frame into data matrix (helper function)
}
\keyword{internal}

