% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConcatScatter.R
\name{DataConcatScatter}
\alias{DataConcatScatter}
\title{Data Concatenated and Scatter Matrix}
\usage{
DataConcatScatter(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{other options}
}
\description{
Data Concatenated and Scatter Matrix
}
\keyword{internal}

