% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataConcatScatter.R
\name{DataConcatScatter}
\alias{DataConcatScatter}
\alias{DataConcatScatter.resample}
\title{Data Concatenated and Scatter Matrix}
\usage{
DataConcatScatter(x, ...)

\method{DataConcatScatter}{resample}(x, targetDim, ...)
}
\arguments{
\item{x}{data}

\item{...}{other options}

\item{targetDim}{target dimension to reduce the data to}
}
\description{
Data Concatenated and Scatter Matrix
}
\keyword{internal}
