% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataToListMatrix.R
\name{dataToListMatrix}
\alias{dataToListMatrix}
\title{Data Conversion to A List of Matrices}
\usage{
dataToListMatrix(x, ...)
}
\arguments{
\item{x}{data not in a list of matrices form}

\item{...}{other options that are not used at the moment}

\item{group}{The grouping variable in a data.frame in quotations}
}
\description{
Data Conversion to A List of Matrices
}
\keyword{internal}
