% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Original.R
\name{Original}
\alias{Original}
\title{Original Data}
\usage{
Original(x, group, targetDim, svdMethod)
}
\arguments{
\item{x}{data}

\item{group}{grouping varible}

\item{targetDim}{target dimension}

\item{svdMethod}{svd function}
}
\description{
Original Data
}
\keyword{internal}

