% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce.R
\name{reduce}
\alias{reduce}
\alias{reduce.list}
\alias{reduce.default}
\title{Reduction}
\usage{
reduce(x, method, targetDim, ...)

\method{reduce}{list}(x, method = SYS, targetDim, ...)

\method{reduce}{default}(x, method = SYS, targetDim, ..., group)
}
\arguments{
\item{x}{a list of matrices or a projection_matrix class object}

\item{method}{a function for a linear dimension reduction method}

\item{targetDim}{a function or numeric that supplies the target dimension}

\item{...}{options past to linear dimension reduction and target dimension methods}

\item{group}{variable in the data the specifies the groupings of the data}
}
\description{
Given a linear reduction method reduction will produce a data reduction
matrix.
}
