% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SConcat.R
\name{SConcat}
\alias{SConcat}
\alias{SConcat.resample}
\title{Sample Covariances Concatenated}
\usage{
SConcat(x, ...)

\method{SConcat}{resample}(x, targetDim, ...)
}
\arguments{
\item{x}{data}

\item{...}{other options}

\item{targetDim}{target dimension to reduce the data to}
}
\description{
Sample Covariances Concatenated
}
\keyword{internal}
