% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SConcat.R
\name{SConcat}
\alias{SConcat}
\title{Sample Covariances Concatenated}
\usage{
SConcat(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{}
}
\description{
Sample Covariances Concatenated
}
\examples{
SConcat(iris, group = Species, targetDim = 1)
}

