% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDiff.R
\name{SDiff}
\alias{SDiff}
\title{Sample Covariance Differences Concatenated}
\usage{
SDiff(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{other variables}
}
\description{
Sample Covariance Differences Concatenated
}
\keyword{internal}

