% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDiff.R
\name{SDiff}
\alias{SDiff}
\title{Sample Covariances Concatenated}
\usage{
SDiff(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{}
}
\description{
Sample Covariances Concatenated
}
\examples{
SDiff(iris, group = Species, targetDim = 1)
}

