% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIR.R
\name{SIR}
\alias{SIR}
\title{Sliced Inverse Regression}
\usage{
SIR(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{}
}
\description{
Sliced Inverse Regression
}
\examples{
SIR(iris, group = Species, targetDim = 1)
}

