% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SYS.R
\name{SYS}
\alias{SYS}
\title{SYS}
\usage{
SYS(x, svdMethod = svd, covEst = Haff_shrinkage, ...)
}
\arguments{
\item{x}{data}

\item{svdMethod}{svd function used for dimesion reduction by default
svd in base is used}

\item{covEst}{covariance estimator used forming M matrix}

\item{...}{options passed to covEst and svd methods}
}
\value{
list of reduced data, projection matrix,
group variable, discrimination function,
m matrix.
}
\description{
SYS
}
